MC-181 : Clean Configurable processor and reader
