+++ Processor/AbstractProductProcessor.php
@@ -598,21 +598,7 @@ abstract class AbstractProductProcessor extends AbstractProcessor
                         'help'  => 'pim_magento_connector.export.skuFirst.help',
                         'label' => 'pim_magento_connector.export.skuFirst.label',
                     ],
-                ],
-                'urlKey' => [
-                    'type'    => 'checkbox',
-                    'options' => [
-                        'help'  => 'pim_magento_connector.export.urlKey.help',
-                        'label' => 'pim_magento_connector.export.urlKey.label',
-                    ],
-                ],
-                'skuFirst' => [
-                    'type'    => 'checkbox',
-                    'options' => [
-                        'help'  => 'pim_magento_connector.export.skuFirst.help',
-                        'label' => 'pim_magento_connector.export.skuFirst.label',
-                    ],
-                ],
+                ]
             ],
             $this->categoryMappingMerger->getConfigurationField(),
             $this->attributeMappingMerger->getConfigurationField()
