+++ Processor/ConfigurableProcessor.php
@@ -2,8 +2,12 @@
 
 namespace Pim\Bundle\MagentoConnectorBundle\Processor;
 
-use Pim\Bundle\CatalogBundle\Manager\ChannelManager;
 use Akeneo\Bundle\BatchBundle\Item\InvalidItemException;
+use Doctrine\Common\Collections\Collection;
+use Pim\Bundle\CatalogBundle\Entity\Channel;
+use Pim\Bundle\CatalogBundle\Manager\ChannelManager;
+use Pim\Bundle\CatalogBundle\Model\ProductInterface;
+use Pim\Bundle\MagentoConnectorBundle\Manager\AssociationTypeManager;
 use Pim\Bundle\MagentoConnectorBundle\Webservice\Webservice;
 use Pim\Bundle\MagentoConnectorBundle\Manager\PriceMappingManager;
 use Pim\Bundle\MagentoConnectorBundle\Guesser\WebserviceGuesser;
@@ -19,21 +23,26 @@ use Pim\Bundle\MagentoConnectorBundle\Webservice\MagentoSoapClientParametersRegi
 /**
  * Magento configurable processor
  *
- * @author    Julien Sanchez <julien@akeneo.com>
- * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
+ * @author    Willy Mesnage <willy.mesnage@akeneo.com>
+ * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
 class ConfigurableProcessor extends AbstractProductProcessor
 {
-    /**
-     * @var ConfigurableNormalizer
-     */
+    /** @var \Pim\Bundle\MagentoConnectorBundle\Normalizer\ConfigurableNormalizer */
     protected $configurableNormalizer;
 
-    /**
-     * @var GroupManager
-     */
+    /** @var AssociationTypeManager */
+    protected $associationTypeManager;
+
+    /** @var GroupManager */
     protected $groupManager;
+
+    /** @var array */
+    protected $processedIds = [];
+
+    /** @var string */
+    protected $pimGrouped;
 
     /**
      * @param WebserviceGuesser                   $webserviceGuesser
@@ -44,8 +53,10 @@ class ConfigurableProcessor extends AbstractProductProcessor
      * @param ChannelManager                      $channelManager
      * @param MagentoMappingMerger                $categoryMappingMerger
      * @param MagentoMappingMerger                $attributeMappingMerger
-     * @param GroupManager                        $groupManager
      * @param MagentoSoapClientParametersRegistry $clientParametersRegistry
+     * @param AttributeManager                    $attributeManager
+     * @param AssociationTypeManager              $associationTypeManager
+     * @param GroupManager                        $groupManager
      */
     public function __construct(
         WebserviceGuesser $webserviceGuesser,
@@ -56,9 +67,10 @@ class ConfigurableProcessor extends AbstractProductProcessor
         ChannelManager $channelManager,
         MagentoMappingMerger $categoryMappingMerger,
         MagentoMappingMerger $attributeMappingMerger,
-        GroupManager $groupManager,
         MagentoSoapClientParametersRegistry $clientParametersRegistry,
-        AttributeManager $attributeManager
+        AttributeManager $attributeManager,
+        AssociationTypeManager $associationTypeManager,
+        GroupManager $groupManager
     ) {
         parent::__construct(
             $webserviceGuesser,
@@ -73,23 +85,28 @@ class ConfigurableProcessor extends AbstractProductProcessor
             $attributeManager
         );
 
-        $this->groupManager = $groupManager;
+        $this->associationTypeManager = $associationTypeManager;
+        $this->groupManager           = $groupManager;
     }
 
     /**
-     * Function called before all process
+     * @return string
      */
-    protected function beforeExecute()
+    public function getPimGrouped()
     {
-        parent::beforeExecute();
+        return $this->pimGrouped;
+    }
 
-        $priceMappingManager          = new PriceMappingManager($this->defaultLocale, $this->currency, $this->channel);
-        $this->configurableNormalizer = $this->normalizerGuesser->getConfigurableNormalizer(
-            $this->getClientParameters(),
-            $this->productNormalizer,
-            $priceMappingManager,
-            $this->visibility
-        );
+    /**
+     * @param string $pimGrouped
+     *
+     * @return ConfigurableProcessor
+     */
+    public function setPimGrouped($pimGrouped)
+    {
+        $this->pimGrouped = $pimGrouped;
+
+        return $this;
     }
 
     /**
@@ -113,30 +130,9 @@ class ConfigurableProcessor extends AbstractProductProcessor
             }
 
             foreach ($configurables as $configurable) {
-                if (empty($configurable['products'])) {
-                    $this->addWarning('The variant group is not associated to any products', [], $configurable);
-                }
-
-                if ($this->magentoConfigurableExist($configurable, $magentoConfigurables)) {
-                    $context = array_merge(
-                        $this->globalContext,
-                        ['attributeSetId' => 0, 'create' => false]
-                    );
-                } else {
-                    $groupFamily = $this->getGroupFamily($configurable);
-                    $context     = array_merge(
-                        $this->globalContext,
-                        [
-                            'attributeSetId' => $this->getAttributeSetId($groupFamily->getCode(), $configurable),
-                            'create'         => true
-                        ]
-                    );
-                }
-
-                try {
-                    $processedItems[] = $this->normalizeConfigurable($configurable, $context);
-                } catch (\Exception $e) {
-                    $this->addWarning($e->getMessage(), [], $configurable['group']);
+                $processedConfigurable = $this->processConfigurable($configurable, $magentoConfigurables);
+                if ($processedConfigurable !== null) {
+                    $processedItems[] = $processedConfigurable;
                 }
             }
         }
@@ -145,40 +141,100 @@ class ConfigurableProcessor extends AbstractProductProcessor
     }
 
     /**
+     * Function called before all process
+     */
+    protected function beforeExecute()
+    {
+        parent::beforeExecute();
+
+        $this->globalContext['pimGrouped'] = $this->pimGrouped;
+        $priceMappingManager               = new PriceMappingManager(
+            $this->defaultLocale,
+            $this->currency,
+            $this->channel
+        );
+        $this->configurableNormalizer      = $this->normalizerGuesser->getConfigurableNormalizer(
+            $this->getClientParameters(),
+            $this->productNormalizer,
+            $priceMappingManager,
+            $this->visibility
+        );
+    }
+
+    /**
+     * Processes configurables
+     *
+     * @param array $configurable
+     * @param array $magentoConfigurables
+     *
+     * @return array|null
+     */
+    protected function processConfigurable(array $configurable, array $magentoConfigurables)
+    {
+        if (empty($configurable['products'])) {
+            $this->addWarning(
+                'The variant group is not associated to any products or product has already been send.',
+                [],
+                $configurable
+            );
+            return;
+        }
+
+        if ($this->magentoConfigurableExist($configurable, $magentoConfigurables)) {
+            $context = array_merge($this->globalContext, ['attributeSetId' => 0, 'create' => false]);
+        } else {
+            $groupFamily = $this->getGroupFamily($configurable);
+            $context     = array_merge(
+                $this->globalContext,
+                [
+                    'attributeSetId' => $this->getAttributeSetId($groupFamily->getCode(), $configurable),
+                    'create'         => true
+                ]
+            );
+        }
+
+        try {
+            $normalizedConfigurable = $this->normalizeConfigurable($configurable, $context);
+        } catch (\Exception $e) {
+            $this->addWarning($e->getMessage(), [], $configurable);
+            return;
+        }
+
+        return $normalizedConfigurable;
+    }
+
+    /**
      * Normalize the given configurable
      *
-     * @param array $configurable The given configurable
-     * @param array $context      The context
+     * @param array $configurable
+     * @param array $context
      *
-     * @throws InvalidItemException If a normalization error occured
-     * @return array                processed item
+     * @return array
      */
-    protected function normalizeConfigurable($configurable, $context)
+    protected function normalizeConfigurable(array $configurable, array $context)
     {
-        $processedItem = $this->configurableNormalizer->normalize(
+        return $this->configurableNormalizer->normalize(
             $configurable,
             AbstractNormalizer::MAGENTO_FORMAT,
             $context
         );
-
-        return $processedItem;
     }
 
     /**
-     * Test if a configurable allready exist on magento platform
+     * Test if a configurable already exist on magento platform
      *
      * @param array $configurable         The configurable
      * @param array $magentoConfigurables Magento configurables
      *
      * @return bool
      */
-    protected function magentoConfigurableExist($configurable, $magentoConfigurables)
+    protected function magentoConfigurableExist(array $configurable, array $magentoConfigurables)
     {
         foreach ($magentoConfigurables as $magentoConfigurable) {
             if ($magentoConfigurable['sku'] == sprintf(
-                Webservice::CONFIGURABLE_IDENTIFIER_PATTERN,
-                $configurable['group']->getCode()
-            )) {
+                    Webservice::CONFIGURABLE_IDENTIFIER_PATTERN,
+                    $configurable['group']->getCode()
+                )) {
                 return true;
             }
         }
@@ -191,11 +247,9 @@ class ConfigurableProcessor extends AbstractProductProcessor
      *
      * @param array $configurable
      *
-     * @throws InvalidItemException If there are two products with different families
-     *
-     * @return Family
+     * @return \Pim\Bundle\CatalogBundle\Entity\Family;
      */
-    protected function getGroupFamily($configurable)
+    protected function getGroupFamily(array $configurable)
     {
         $groupFamily = $configurable['products'][0]->getFamily();
 
@@ -223,13 +277,20 @@ class ConfigurableProcessor extends AbstractProductProcessor
      */
     protected function getProductsForGroups(array $products, array $groupsIds)
     {
-        $groups = [];
+        $channel = $this->channelManager->getChannelByCode($this->getChannel());
+        $groups  = [];
 
         foreach ($products as $product) {
             foreach ($product->getGroups() as $group) {
                 $groupId = $group->getId();
 
                 if (in_array($groupId, $groupsIds)) {
+                    $groupProducts = $group->getProducts();
+                    $exportableProducts = $this->getExportableProducts(
+                        $channel,
+                        $groupProducts
+                    );
+
                     if (!isset($groups[$groupId])) {
                         $groups[$groupId] = [
                             'group'    => $group,
@@ -237,12 +298,92 @@ class ConfigurableProcessor extends AbstractProductProcessor
                         ];
                     }
 
-                    $groups[$groupId]['products'][] = $product;
+                    if (!isset($this->processedIds[$groupId])) {
+                        $this->processedIds[$groupId] = [];
+                    }
+
+                    foreach ($exportableProducts as $exportableProduct) {
+                        $exportableProductId = $exportableProduct->getId();
+                        if (!in_array($exportableProductId, $this->processedIds[$groupId])) {
+                            $groups[$groupId]['products'][] = $exportableProduct;
+                            $this->processedIds[$groupId][] = $exportableProductId;
+                        }
+                    }
                 }
             }
         }
 
         return $groups;
+    }
+
+    /**
+     *
+     * Returns ready to export variant group products
+     *
+     * @param Channel          $channel
+     * @param Collection|array $groupProducts
+     *
+     * @return array
+     */
+    protected function getExportableProducts(Channel $channel, $groupProducts)
+    {
+        $exportableProducts = [];
+        $rootCategoryId     = $channel->getCategory()->getId();
+
+        foreach ($groupProducts as $product) {
+            $productCategories = $product->getCategories()->toArray();
+            if ($this->isProductComplete($product, $channel) &&
+                false !== $productCategories &&
+                $this->doesProductBelongToChannel($productCategories, $rootCategoryId)
+            ) {
+                $exportableProducts[] = $product;
+            }
+        }
+
+        return $exportableProducts;
+
+    }
+
+    /**
+     * Is the given product complete for the given channel?
+     *
+     * @param ProductInterface $product
+     * @param Channel $channel
+     *
+     * @return bool
+     */
+    protected function isProductComplete(ProductInterface $product, Channel $channel)
+    {
+        $completenesses = $product->getCompletenesses()->toArray();
+        foreach ($completenesses as $completeness) {
+            if ($completeness->getChannel()->getId() === $channel->getId() &&
+                $completeness->getRatio() < 100
+            ) {
+                return false;
+            }
+        }
+
+        return true;
+    }
+
+    /**
+     * Compute the belonging of a product to a channel.
+     * Validating one of its categories has the same root as the channel root category.
+     *
+     * @param \ArrayIterator|array $productCategories
+     * @param int                  $rootCategoryId
+     *
+     * @return bool
+     */
+    protected function doesProductBelongToChannel($productCategories, $rootCategoryId)
+    {
+        foreach ($productCategories as $category) {
+            if ($category->getRoot() === $rootCategoryId) {
+                return true;
+            }
+        }
+
+        return false;
     }
 
     /**
@@ -253,5 +394,28 @@ class ConfigurableProcessor extends AbstractProductProcessor
     protected function getGroupRepository()
     {
         return $this->groupManager->getRepository();
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getConfigurationFields()
+    {
+        return array_merge(
+            parent::getConfigurationFields(),
+            [
+                'pimGrouped' => [
+                    'type'    => 'choice',
+                    'options' => [
+                        'choices' => $this->associationTypeManager->getAssociationTypeChoices(),
+                        'help'    => 'pim_magento_connector.export.pimGrouped.help',
+                        'label'   => 'pim_magento_connector.export.pimGrouped.label',
+                        'attr' => [
+                            'class' => 'select2',
+                        ],
+                    ],
+                ]
+            ]
+        );
     }
 }
