+++ Resources/config/processors.yml
@@ -1,7 +1,6 @@
 parameters:
     pim_magento_connector.processor.product.class:             Pim\Bundle\MagentoConnectorBundle\Processor\ProductProcessor
     pim_magento_connector.processor.configurable.class:        Pim\Bundle\MagentoConnectorBundle\Processor\ConfigurableProcessor
-    pim_magento_connector.processor.delta_configurable.class:  Pim\Bundle\MagentoConnectorBundle\Processor\DeltaConfigurableProcessor
     pim_magento_connector.processor.product_association.class: Pim\Bundle\MagentoConnectorBundle\Processor\ProductAssociationProcessor
     pim_magento_connector.processor.category.class:            Pim\Bundle\MagentoConnectorBundle\Processor\CategoryProcessor
     pim_magento_connector.processor.attribute.class:           Pim\Bundle\MagentoConnectorBundle\Processor\AttributeProcessor
@@ -39,25 +38,10 @@ services:
             - '@pim_catalog.manager.channel'
             - '@pim_magento_connector.merger.category_mapping'
             - '@pim_magento_connector.merger.attribute_code_mapping'
-            - '@pim_magento_connector.manager.group'
             - '@pim_magento_connector.webservice.magento_soap_client_parameters_registry'
             - '@pim_magento_connector.manager.attribute'
-
-    pim_magento_connector.processor.delta_configurable:
-        class: %pim_magento_connector.processor.delta_configurable.class%
-        parent: pim_magento_connector.item.magento_item_step
-        arguments:
-            - '@pim_magento_connector.guesser.magento_webservice'
-            - '@pim_magento_connector.guesser.magento_normalizer'
-            - '@pim_magento_connector.manager.locale'
-            - '@pim_magento_connector.merger.storeview_mapping'
-            - '@pim_magento_connector.manager.currency'
-            - '@pim_catalog.manager.channel'
-            - '@pim_magento_connector.merger.category_mapping'
-            - '@pim_magento_connector.merger.attribute_code_mapping'
+            - '@pim_magento_connector.manager.association_type'
             - '@pim_magento_connector.manager.group'
-            - '@pim_magento_connector.webservice.magento_soap_client_parameters_registry'
-            - '@pim_magento_connector.manager.attribute'
 
     pim_magento_connector.processor.product_association:
         class: %pim_magento_connector.processor.product_association.class%
